+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+
 ## 1.1.5 (2023-12-18)
 - School Assembly page changes
   - JSON file now saves with new line formatting for human readability
   - Loading the right links for default 2 videos fixed.
+- STEP Events page updates
+  - Additional files section below video for download
+  - Editing function for created events
 
 ## 1.1.4 (2023-12-06)
 - STEP Events page updates